fix(web): `getQueryClient` 함수 분리
